+++ examples/native/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text } from 'react-native';
-import { Form } from '@jbk/react-form/native';
+import { Form } from '@jbknowledge/react-form/native';
 
 import Input from './Input';